+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Only the Dart 2 dev SDK (`>=2.0.0-dev.22.0`) is now supported.
 * Add support for `Map.{addEntries|entries}` for Dart 2.0.
+* Add support for `Iterable|List|Map|Set|Queue.{cast|retype}` for Dart 2.0.
 
 ## 1.14.6